fix bug with applying reactions
the bug caused not all of the reactions aplicable to a tp_hybrid
to be applied, as well as causing many of the resulting products
to be invalid molecules. The fix was to change how the reactions
were hashed to use parent_side_chain _ids and match based on
connection atom index, as well as separate pictet and template
dependent reactions from each other.