update reaction structure, add reaction
change IUniMolecularReaction to AbstractUniMolecularReaction
and IBiMolecularReaction to AbstractBiMolecularReaction. Remove
IDisJointReaction and IJointReaction classes and add
AbstractDMSideChainReaction and AbstractBMMonomerReaction. These
classes now represent reactions that are bimolecular and have
either a sidechain or monomer as a reactant. However the monomer
must be specifically created in the reaction, hence it was given
its own class.

Added PyrroloIndolene reaction class