implement PeptidePublicationPlanner
PeptidePublicationPlanner generates tuples of monomer indexes to
be used to generate the peptides used in the library. It ensures
all monomers aree used atleast once at each position of a peptide.

Also implemented an IO class to handle the loading at writing of
PeptidePublicationPlanner data and helper function
generate_peptide_plan() in utils.