implement way to use reduced sidechain set
In case user wants to use a reduced set of sidechains to generate
the library, the code now allows for loading a sdf file containing
those sidechains and filtering out monomers that dont have
sidechains within that set. Peptide planning can then be done with
this reduced set. Makes it much easier than reloading and whole new
set with new ids and matching thoes to the regiosqm and pka
predictions